fix(sql queries): fix small things
- fix values name in INSERT query
- fix column type in users table

+++ database/queries/users_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS users (
     id       SERIAL,
-    userid   int,
+    userid   bigint,
     username varchar(64),
     lang     char(2),
     UNIQUE   (userid)
